mingw support (#802)
Allow for compiling with mingw, and for cross-compilation support via the included docker container for convenience.

As mingw uses gcc, some changes to the client's codebase were required to fix compilation there, as it's a little more strict than MSVC. This should reduce the potential changes for when we actually support compilation with gcc down the track.

Some changes were also made to timeout checks on thread ticks; with mingw's implementation (not sure if this extends to gcc), threads were often unable to timeout due to spurious wakeups interrupting them frequently enough that an entire timeout could never complete.

We also make sure to only build N3Base_client with OPENKO_BUILD_CLIENT enabled, and N3Base_tool with OPENKO_BUILD_TOOLS, rather than having them both built when either is specified. This is because it doesn't support MFC, so we cannot build tools with mingw -- only the client.

Finally, @kenner2 adds a Docker container to make cross-compilation simple, though this will probably just be removed again later.

* Support building with mingw

This mostly covers our bases, just leaving libjpeg-turbo as the perpetual issue that it is, and more than likely still requiring some build-related config to ensure everything is statically linked still (aside from DX9).

* Import DX9 libs as needed

* Statically link to everything for consistency across projects.

* feat: Utility to cross-compile and run client in Linux (#797)

* Only build N3Base_client or N3Base_tool if we're building the client or tools respectively

Before we built both regardless, if we were building either.

* More robustly detect & ignore spurious thread wakeups

* Add algorithm import for std::clamp support

---------

+++ src/Client/WarFare/UIHotKeyDlg.cpp
@@ -1052,7 +1052,7 @@ void CUIHotKeyDlg::RenderCooldown(const __IconItemSkill* pSkill, float fCooldown
 	const float halfWidth    = static_cast<float>(centerX - rc.left);
 	const float halfHeight   = static_cast<float>(centerY - rc.top);
 
-	const float radius       = std::sqrtf(std::pow(halfWidth, 2.0f) + std::pow(halfHeight, 2.0f));
+	const float radius       = std::sqrt(std::pow(halfWidth, 2.0f) + std::pow(halfHeight, 2.0f));
 
 	float progress           = 0.0f;
